+++ opal/core/referencedata/__init__.py
@@ -1,0 +1,3 @@
+"""
+Opal core referencedata package
+"""