+++ opal/core/referencedata/plugin.py
@@ -1,0 +1,10 @@
+"""
+Plugin definition for opal.core.referencedata
+"""
+from opal.core import plugins
+
+
+class ReferencedataPlugin(plugins.OpalPlugin):
+    """
+    The plugin entrypoint for core referencedata
+    """